+++ src/database.ts
@@ -9,13 +9,15 @@ import * as path from "path";
 import { DB_DIR_NAME, DB_FILE_NAME, BACKUP_DIR_NAME } from "./constants.js";
 import { runMigrations } from "./migrations.js";
 import { createRepositories, type Repositories } from "./repositories/index.js";
-import { CompactionService, ProjectScanService, GitService, EventTriggerService } from "./services/index.js";
+import { CompactionService, ProjectScanService, GitService, EventTriggerService, UpdateService } from "./services/index.js";
+import { SERVER_VERSION } from "./constants.js";
 
 export interface Services {
   compaction: CompactionService;
   scan: ProjectScanService;
   git: GitService;
   events: EventTriggerService;
+  update: UpdateService;
 }
 
 let _db: DatabaseType | null = null;
@@ -54,6 +56,7 @@ export async function initDatabase(projectRoot: string): Promise<DatabaseType> {
     scan: new ProjectScanService(_repos),
     git: new GitService(projectRoot),
     events: new EventTriggerService(_repos),
+    update: new UpdateService(_repos, SERVER_VERSION),
   };
 
   return _db;